Remove SimpleMenuDialog.
Instead we position a plain PlasmaComponents.Dialog, relying on
QQuickItem::mapToGlobal from Qt 5.7 to factor in the screen
coordinates of the CompactRepresentation item.

This isn't perfect yet - it can't take struts on a perpendicular
screen edge into account, because applets only get
plasmoid.screenGeometry to work with. The superior
plasmoid.availableScreenRect is only available to containments.

CCMAIL:notmart@gmail.com