Fix favorites on 5.11
Synced to code from Kicker, but falling back to the old code if we
don't find initForClient in the favoritesModel. Hopefully this way
it continues to work on 5.10 too.

Time for an emergency 1.0.3 release - somewhat unhappy after I
asked to verify this wouldn't be needed for Store applets.

CCMAIL:ivan.cukic@kde.org