Correcting popup position issues with dual screen
If you have two screens, and you place Screen A above Screen B, and then you place the simple menu widget on a panel at the top of screen B (which is below screen A), the popup shows up on the top of screen A. 

Same bugs appears every time you place your panel with the widget "between" both screens. Adding screenGeom fixed it for me.